@@ -1,10 +1,12 @@
 package com.davistiba.wedemyserver.config;
 
 import com.davistiba.wedemyserver.dto.UserDTO;
-import com.davistiba.wedemyserver.models.User;
 import com.davistiba.wedemyserver.service.MyUserDetailsService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -14,9 +16,6 @@ import org.springframework.security.oauth2.core.oidc.user.OidcUser;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
@@ -44,15 +43,14 @@ public class CustomAuthSuccessHandler implements AuthenticationSuccessHandler {
     public void onAuthenticationSuccess(HttpServletRequest request,
                                         HttpServletResponse response,
                                         Authentication auth) throws IOException, ServletException {
-        if (auth.getPrincipal() instanceof OidcUser) {
+        if (auth.getPrincipal() instanceof OidcUser oidcUser) {
             //if Google Login
-            OidcUser oidcUser = (OidcUser) auth.getPrincipal();
             myUserDetailsService.processOAuthPostLogin(oidcUser, request.getSession());
             response.sendRedirect(FRONTEND_URL);
             return;
         }
-        User loggedInUser = (User) auth.getPrincipal();
-        UserDTO userInfo = modelMapper.map(loggedInUser, UserDTO.class);
+        MainUserDetails loggedInUser = (MainUserDetails) auth.getPrincipal();
+        UserDTO userInfo = modelMapper.map(loggedInUser.getUser(), UserDTO.class);
 
         request.getSession().setAttribute(MyUserDetailsService.USERID, userInfo.getId());
         Map<String, Object> authResponse = new HashMap<>();
