@@ -5,6 +5,8 @@ import com.davistiba.wedemyserver.dto.SalesDTO;
 import com.davistiba.wedemyserver.repository.OrderItemRepository;
 import com.davistiba.wedemyserver.repository.SalesRepository;
 import com.davistiba.wedemyserver.service.MyUserDetailsService;
+import jakarta.servlet.http.HttpSession;
+import jakarta.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -12,9 +14,6 @@ import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
-
-import javax.servlet.http.HttpSession;
-import javax.validation.constraints.NotNull;
 
 @RestController
 @RequestMapping(path = "/sales", produces = MediaType.APPLICATION_JSON_VALUE)
