@@ -5,13 +5,13 @@ import com.davistiba.wedemyserver.models.Review;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.PagingAndSortingRepository;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface ReviewRepository extends PagingAndSortingRepository<Review, Integer> {
+public interface ReviewRepository extends CrudRepository<Review, Integer> {
 
     @Query("SELECT r FROM Review r WHERE r.user.id = ?1 AND r.course.id = ?2")
     Optional<Review> findByUserIdAndCourseId(Integer userId, Integer courseId);
