@@ -1,10 +1,13 @@
 package com.davistiba.wedemyserver.service;
 
+import com.davistiba.wedemyserver.config.MainUserDetails;
 import com.davistiba.wedemyserver.models.AuthProvider;
 import com.davistiba.wedemyserver.models.CustomOAuthUser;
 import com.davistiba.wedemyserver.models.User;
 import com.davistiba.wedemyserver.models.UserRole;
 import com.davistiba.wedemyserver.repository.UserRepository;
+import jakarta.servlet.http.HttpSession;
+import jakarta.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -12,8 +15,6 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.oauth2.core.oidc.user.OidcUser;
 import org.springframework.stereotype.Service;
 
-import javax.servlet.http.HttpSession;
-import javax.validation.constraints.NotNull;
 import java.util.Optional;
 
 
@@ -27,8 +28,9 @@ public class MyUserDetailsService implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return userRepository.findByEmail(username)
+        var user = userRepository.findByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException("Not found"));
+        return new MainUserDetails(user);
     }
 
     /**
@@ -47,6 +49,7 @@ public class MyUserDetailsService implements UserDetailsService {
             newUser.setFullname(m.getName());
             newUser.setEmail(m.getEmail());
             newUser.setConfirmPass("WHATEVER!"); //<-- anything, but not NULL
+            newUser.setEnabled(true);
             newUser.setAuthProvider(AuthProvider.GOOGLE);
             newUser.setUserRole(UserRole.ROLE_STUDENT);
 
